requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
python-engineio==3.12.1
python-socketio==4.5.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
websocket-client==0.57.0
eventlet==0.25.2
